Send size stat to double

+++ include/stat.h
@@ -10,11 +10,11 @@ typedef struct statistics {
     int32_t request_parsing_error;
 
     int32_t nat_sended_to_client_error;
-    int32_t nat_sended_to_client_size;
+    double nat_sended_to_client_size;
     int32_t nat_sended_to_client;
 
     int32_t nat_sended_to_dev_error;
-    int32_t nat_sended_to_dev_size;
+    double nat_sended_to_dev_size;
     int32_t nat_sended_to_dev;
 
     int32_t latency_sended_to_dev_sum;
